✨ Added RP select context menu - APIにレジストする関数を整理
