+++ src/lib/commandUtils.ts
@@ -6,89 +6,6 @@ import fs from "fs";
 import path from "path";
 
 /**
- * @param {Client} client
- */
-export const addCommand = async (client: Client) => {
-  console.log(`\u001b[32m===Pushing Command Data===\u001b[0m`);
-  const config = client.config;
-  const token = config.token;
-  const rest = new REST({ version: "10" }).setToken(token);
-
-  const testGuild = config.dev.testGuild;
-
-  let command_int = 0;
-  const globalCommands = [] as RESTPostAPIChatInputApplicationCommandsJSONBody[];
-  const adminGuildCommands = [] as RESTPostAPIChatInputApplicationCommandsJSONBody[];
-  const commandFolders = fs.readdirSync(path.resolve(__dirname, `../executors/chatInputCommands`));
-
-  function cmdToArray(
-    array: RESTPostAPIChatInputApplicationCommandsJSONBody[],
-    command: ChatInputCommand,
-    file: string,
-    notice = ""
-  ) {
-    try {
-      array.push((command.data as SlashCommandBuilder).toJSON());
-      command_int++;
-      console.log(`${notice} ${file} has been added.`);
-    } catch (error) {
-      console.error(`${notice} An Error Occured in ${file} \nエラー内容\n ${error}`);
-    }
-  }
-
-  async function putToDiscord(
-    array: RESTPostAPIChatInputApplicationCommandsJSONBody[],
-    guild: undefined | string = undefined
-  ) {
-    if (guild) {
-      await rest.put(Routes.applicationGuildCommands(client.application?.id as string, guild), {
-        body: array,
-      });
-    } else {
-      await rest.put(Routes.applicationCommands(client.application?.id as string), {
-        body: array,
-      });
-    }
-  }
-
-  for (const folder of commandFolders) {
-    console.log(`[Init]Adding ${folder} commands...`);
-    const commandFiles = fs
-      .readdirSync(path.resolve(__dirname, `../executors/chatInputCommands/${folder}`))
-      .filter((file) => file.endsWith(".js") || (file.endsWith(".ts") && !file.endsWith(".d.ts")));
-    for (const file of commandFiles) {
-      const command = require(path.resolve(
-        __dirname,
-        `../executors/chatInputCommands/${folder}/${file}`
-      )) as ChatInputCommand;
-      if (command.adminGuildOnly) {
-        cmdToArray(adminGuildCommands, command, file, "[Admin]");
-        continue;
-      }
-      //if (command.onlyCommand) continue;
-      cmdToArray(globalCommands, command, file, "[Global]");
-    }
-    console.log(`[Init]${folder} added.`);
-  }
-
-  try {
-    console.log(`[Init]Registering ${command_int}...`);
-
-    //Admin
-    putToDiscord(adminGuildCommands, testGuild);
-    console.log(`[Init]Registered Admin Guild Slash Commands.`);
-
-    //Global
-    putToDiscord(globalCommands);
-    console.log(`[Init]Registered Global Slash Commands.`);
-
-    console.log(`[Init]Registered All Slash Commands.`);
-  } catch (error) {
-    console.error("[error]", error);
-  }
-};
-
-/**
  * Adds subcommands to the provided data object.
  *
  * @param {string} name - The name of the command.
@@ -148,7 +65,6 @@ export const subCommandHandling = (name: string) => {
 };
 
 export default {
-  addCommand,
   addSubCommand,
   subCommandHandling,
 };
